tables_to_genies/app/requirements.txt
# Databricks SDK and SQL
databricks-sdk>=0.20.0
databricks-sql-connector>=3.0.0

# Graph visualization
dash-cytoscape>=1.0.2

# GraphRAG
graphrag>=0.3.0

# Data generation (for testing)
faker>=24.0.0

# Layout
dash-bootstrap-components>=1.5.0
